+++ package.json
@@ -7,15 +7,17 @@
     "dev": "vite",
     "build": "vite build",
     "preview": "vite preview",
-    "check": "svelte-check --tsconfig ./tsconfig.json"
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json"
   },
   "devDependencies": {
+    "@sveltejs/adapter-auto": "^3.0.0",
     "@sveltejs/vite-plugin-svelte": "^3.0.1",
     "@tsconfig/svelte": "^5.0.2",
     "@types/lodash": "^4.14.202",
     "@zerodevx/svelte-toast": "^0.9.5",
     "autoprefixer": "^10.4.16",
     "postcss": "^8.4.33",
+    "@sveltejs/kit": "^2.0.0",
     "svelte": "^4.2.8",
     "svelte-check": "^3.6.2",
     "svelte-hamburgers": "^4.2.0",
