+++ src/lib/Gallery/svelte-carousel.d.ts
