+++ src/routes/app.css
@@ -6,7 +6,7 @@
   font-family: Inter, system-ui, Avenir, Helvetica, Arial, sans-serif;
   line-height: 1.5;
   font-weight: 400;
-  --background-image: url('./assets/fleshEphemeral.jpg');
+  --background-image: url('$lib/images/fleshEphemeral.jpg');
   color-scheme: light dark;
   color: rgba(255, 255, 255, 0.87);
   font-synthesis: none;
